services/user-service/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-filter==25.1
psycopg==3.2.4
gunicorn==23.0.0
python-decouple==3.8
drf-spectacular==0.28.0
pillow==11.3.0
requests==2.32.4